Continue work on getting some initial functionality.
The program now just prints what it should be doing to simplify
debugging.

Added a RmMonad which is a StateT around IO, deriving both MonadState
and MonadIO. This allows it to keep track of some paths as state, and
also perform the IO-operations required. The state now contains three
paths: The temp-dir where the files are to be moved, the basepath of the
argument passed to the program (~ $ rm.hs a/b/c would give a basepath of
~/a/b) and a relative path which is to be preserved in the temp-dir (if
a contains b and c/d, $ rm.hs a should move a/b and a/c/d to the
temp-dir.

This required some changes both in Main and Rm and is not completely
tested yet.